making parallel

+++ d2q9-bgk.c
@@ -406,12 +406,12 @@ int main(int argc, char* argv[])
   t_speed* output= (t_speed*)malloc(sizeof(t_speed) * (params.ny * params.nx));
   print_halo_fushion(params,*local_cells_ptr,work);
   int temp = rank*2;
-  int array[4];
-  MPI_Gather(&temp,1,MPI_INT,array,1,MPI_INT,0,MPI_COMM_WORLD);
+  float array[4];
+  MPI_Gather(&temp,1,MPI_FLOAT,array,1,MPI_FLOAT,0,MPI_COMM_WORLD);
   if(rank==0){
   printf("\n");
   for(int i = 0;i<2;i++){
-    printf("%d  ",array[i]);
+    printf("%f  ",array[i]);
   }
   printf("\n");
   }
